+++ CHANGELOG.md
@@ -1,8 +1,12 @@
 # Change Log
 
-All notable changes to the "bioviewer" extension will be documented in this file.
+All notable changes to the "BioViewer" extension will be documented in this file.
 
 ## [Unreleased]
+
+## [0.0.3] - 2024-07-30
+
+- Fix bugs in loading files from server
 
 ## [0.0.2] - 2024-07-29